refactor: update all imports to use index file
