+++ src/errors/index.ts
@@ -1,2 +1,3 @@
 export * from './application.error';
 export * from './validation.error';
+export * from './error-messages';
