+++ src/repositories/index.ts
@@ -1,6 +1,5 @@
 export * from './user.repository';
 export * from './customer.repository';
-export * from './branch.repository';
 export * from './menu.repository';
 export * from './order.repository';
 export * from './payment.repository';
